Added patron badges & changed wait what
And fixed Votifier repo
